requirements.txt
selenium
pytest
requests
webdriver-manager
allure-pytest
pytest-xdist
pytest-rerunfailures
python-dotenv
pyyaml
tenacity
faker